+++ DialogTools/src/pwd.cpp
@@ -36,7 +36,7 @@ void DoPwd(HANDLE aDlg)
   Info.SendDlgMessage(aDlg,DM_GETDLGITEMSHORT,itemID,&DialogItem);
   if(DialogItem.Type==DI_PSWEDIT)
   {
-    long length=Info.SendDlgMessage(aDlg,DM_GETTEXT,itemID,0)+1;
+    size_t length=Info.SendDlgMessage(aDlg,DM_GETTEXT,itemID,0)+1;
     TCHAR *buffer=(TCHAR *)HeapAlloc(GetProcessHeap(),HEAP_ZERO_MEMORY,length*sizeof(TCHAR));
     if(buffer)
     {
