+++ DialogTools/src/search.cpp
@@ -36,7 +36,7 @@ void DoSearch(HANDLE aDlg)
   Info.SendDlgMessage(aDlg,DM_GETDLGITEMSHORT,itemID,&DialogItem);
   if(DialogItem.Type==DI_EDIT||DialogItem.Type==DI_FIXEDIT)
   {
-    long length=Info.SendDlgMessage(aDlg,DM_GETTEXT,itemID,0)+1;
+    size_t length=Info.SendDlgMessage(aDlg,DM_GETTEXT,itemID,0)+1;
     TCHAR *buffer=(TCHAR *)HeapAlloc(GetProcessHeap(),HEAP_ZERO_MEMORY,length*sizeof(TCHAR));
     if(buffer)
     {
@@ -52,15 +52,15 @@ void DoSearch(HANDLE aDlg)
       }
       FarDialogItem DialogItems[]=
       {//       Type            X1 Y1 X2 Y2 Se Fl                                     Data */
-       /*00*/ { DI_DOUBLEBOX,   3, 1,55, 9, {0},NULL,                       NULL, DIF_BOXCOLOR,                          GetMsg(mSearchCaption) ,0,0},
-       /*01*/ { DI_TEXT,        5, 2, 0, 2, {0},NULL,                       NULL, 0,                                     GetMsg(mSearch)        ,0,0},
-       /*02*/ { DI_EDIT,        5, 3,53, 3, {0},_T("DialogManager.Search"), NULL, DIF_USELASTHISTORY,                    _T("")                 ,0,0},
-       /*03*/ { DI_TEXT,        0, 4, 0, 4, {0},NULL,                       NULL, DIF_BOXCOLOR|DIF_SEPARATOR,            _T("")                 ,0,0},
-       /*04*/ { DI_CHECKBOX,    5, 5, 0, 5, {c},NULL,                       NULL, 0,                                     GetMsg(mCase)          ,0,0},
-       /*05*/ { DI_CHECKBOX,    5, 6, 0, 6, {p},NULL,                       NULL, 0,                                     GetMsg(mCurPos)        ,0,0},
-       /*06*/ { DI_TEXT,        0, 7, 0, 7, {0},NULL,                       NULL, DIF_BOXCOLOR|DIF_SEPARATOR,            _T("")                 ,0,0},
-       /*07*/ { DI_BUTTON,      0, 8, 0, 8, {0},NULL,                       NULL, DIF_CENTERGROUP|DIF_DEFAULTBUTTON,     GetMsg(mOk)            ,0,0},
-       /*08*/ { DI_BUTTON,      0, 8, 0, 8, {0},NULL,                       NULL, DIF_CENTERGROUP,                       GetMsg(mCancel)        ,0,0}
+       /*00*/ { DI_DOUBLEBOX,   3, 1,55, 9, {0},          NULL,                       NULL, DIF_BOXCOLOR,                          GetMsg(mSearchCaption) ,0,0},
+       /*01*/ { DI_TEXT,        5, 2, 0, 2, {0},          NULL,                       NULL, 0,                                     GetMsg(mSearch)        ,0,0},
+       /*02*/ { DI_EDIT,        5, 3,53, 3, {0},          _T("DialogManager.Search"), NULL, DIF_USELASTHISTORY,                    _T("")                 ,0,0},
+       /*03*/ { DI_TEXT,        0, 4, 0, 4, {0},          NULL,                       NULL, DIF_BOXCOLOR|DIF_SEPARATOR,            _T("")                 ,0,0},
+       /*04*/ { DI_CHECKBOX,    5, 5, 0, 5, {(intptr_t)c},NULL,                       NULL, 0,                                     GetMsg(mCase)          ,0,0},
+       /*05*/ { DI_CHECKBOX,    5, 6, 0, 6, {(intptr_t)p},NULL,                       NULL, 0,                                     GetMsg(mCurPos)        ,0,0},
+       /*06*/ { DI_TEXT,        0, 7, 0, 7, {0},          NULL,                       NULL, DIF_BOXCOLOR|DIF_SEPARATOR,            _T("")                 ,0,0},
+       /*07*/ { DI_BUTTON,      0, 8, 0, 8, {0},          NULL,                       NULL, DIF_CENTERGROUP|DIF_DEFAULTBUTTON,     GetMsg(mOk)            ,0,0},
+       /*08*/ { DI_BUTTON,      0, 8, 0, 8, {0},          NULL,                       NULL, DIF_CENTERGROUP,                       GetMsg(mCancel)        ,0,0}
       };
       CFarDialog dialog;
       int n=dialog.Execute(MainGuid,SearchDialogGuid,-1,-1,59,11,NULL,DialogItems,ArraySize(DialogItems),0,0,SearchDialogProc,NULL);
